Create modal for mobile version
