+++ src/App.css
@@ -24,6 +24,10 @@
   zrobić czystym cssem i trzeba będzie 
   za każdym razem wpisywać media.*/
 
+  --mobile: 320px;
+  --tablet: 772px;
+  --desktop: 1280px;
+
   /*Font Sizes*/
   --title-desktop: 102px;
   --subtitle-mobile: 13px;
@@ -53,5 +57,4 @@
   --border-none: none;
   --border-thin: 1px solid;
   --border-thick: 2px solid;
-  
 }
