+++ src/App.jsx
@@ -7,12 +7,14 @@ import {
   useSearchParams,
 } from "react-router-dom";
 import Home from "./pages/Home/Home";
-import {Report} from "./pages/Report/Report"
+import MainPage from "./pages/MainPage/MainPage";
+import { Report } from "./pages/Report/Report";
 
 const App = () => {
   return (
     <Routes>
       <Route path="/" element={<Home />} />
+      <Route path="/main" element={<MainPage />} />
       <Route path="/report" element={<Report />} />
     </Routes>
   );
