+++ src/components/Background/Background.jsx
@@ -1,0 +1,7 @@
+import css from "./Background.module.css";
+
+const Background = () => {
+  return <div className={css.background}></div>;
+};
+
+export default Background;
