--- src/components/HomePage/MainPage/MainPage.jsx
@@ -1,22 +1,0 @@
-import Balance from "../Balance/Balance";
-import BalanceModal from "../BalanceModal/BalanceModal";
-import MobileHome from "../MobileHome/MobileHome";
-import Navigation from "../Navigation/Navigation";
-import ReportsLink from "../ReportsLink/ReportsLink";
-import Today from "../../Today/Today";
-import ProductForm from "../ProductForm/ProductForm";
-
-const MainPage = () => {
-  return (
-    <>
-      <MobileHome />
-      <ReportsLink />
-      <Balance />
-      <Navigation />
-      {/* <Today /> */}
-      <ProductForm />
-    </>
-  );
-};
-
-export default MainPage;
