+++ src/components/MainContainer/MainContainer.module.css
@@ -3,23 +3,22 @@
   margin: 0 auto;
   padding: 0 20px;
   overflow: hidden;
-  /* background: orange; */
 }
 
-@media (min-width: 320px) {
+@media (min-width: var(--mobile)) {
   .container {
     max-width: 320px;
   }
 }
 
-@media (min-width: 772px) {
+@media (min-width: var(--tablet)) {
   .container {
     max-width: 772px;
     padding: 0 32px;
   }
 }
 
-@media (min-width: 1280px) {
+@media (min-width: var(--desktop)) {
   .container {
     max-width: 1280px;
     padding: 0 16px;
