+++ src/components/MainPage/BalanceModal/BalanceModal.jsx
@@ -6,6 +6,7 @@ const BalanceModal = () => {
       <p className={css.textFirst}>
         Hello! To get started, enter the current balance of your account!
       </p>
+      <br />
       <p className={css.textSecond}>
         You can't spend money until you have it :&#41;
       </p>
