+++ src/components/MainPage/MobileForm/MobileForm.module.css
@@ -1,0 +1,42 @@
+.modal {
+  /* width: 100%;
+  height: 100%; */
+  background: #fff;
+  border: none;
+  padding: 20px;
+  top: 50%;
+  left: 50%;
+  transform: translate(-50%, -50%);
+  width: 100vw;
+  height: 100vh;
+  z-index: 11;
+  transition-timing-function: cubic-bezier(0.4, 0, 0.2, 1);
+}
+
+.modal::backdrop {
+  background: #fff;
+}
+
+.backdrop {
+  /* position: fixed;
+    left: 0;
+    top: 0; */
+  width: 100vw;
+  height: 100vh;
+}
+
+.wrapper {
+  width: 100vw;
+  height: 100vh;
+}
+
+.btn {
+  background: none;
+  border: none;
+}
+
+.btnIcon {
+  width: 24px;
+  height: 24px;
+  fill: var(--brand-color);
+}
